Code/requirements.txt
Flask==2.2.5
mysql-connector-python==8.0.33
pandas==2.2.1
openpyxl==3.1.2